chore(suite-native): cleanup tron send code
